Safari-611.1.10/LayoutTests/fast/forms/past-names-map-should-be-used-only-when-named-item-is-empty.html

<!DOCTYPE html>
<html>
<body>
<form><input type="text" name="foo"><input type="text" name="bar"><input type="text" name="bar"></form>
<script src="../../resources/js-test-pre.js"></script>
<script>

var form = document.querySelector('form');
shouldBe("form['foo']", "form.childNodes[0]");
shouldBe("form['bar'][0]", "form.childNodes[1]");
shouldBe("form['bar'][1]", "form.childNodes[2]");
shouldBe("form.childNodes[0].name = 'bar'; form.childNodes[1].name = form.childNodes[2].name = 'foo'; form['foo'].length", "2");

var successfullyParsed = true;

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
